requirements.txt
selenium 
pillow 
instagrapi 
schedule 